+++ src/egon/data/airflow/dags/pipeline.py
@@ -5,7 +5,15 @@ from airflow.operators.python_operator import PythonOperator
 from airflow.utils.dates import days_ago
 import importlib_resources as resources
 
-from egon.data.airflow.tasks import initdb
+from egon.data.datasets import database
+from egon.data.datasets.data_bundle import DataBundle
+from egon.data.datasets.heat_etrago import HeatEtrago
+from egon.data.datasets.heat_supply import HeatSupply
+from egon.data.datasets.osm import OpenStreetMap
+from egon.data.datasets.mastr import mastr_data_setup
+from egon.data.datasets.re_potential_areas import re_potential_area_setup
+from egon.data.datasets.mv_grid_districts import mv_grid_districts_setup
+from egon.data.datasets.vg250 import Vg250
 from egon.data.processing.zensus_vg250 import (
     zensus_population_inside_germany as zensus_vg250,
 )
@@ -16,26 +24,32 @@ import egon.data.importing.era5 as import_era5
 import egon.data.importing.etrago as etrago
 import egon.data.importing.heat_demand_data as import_hd
 import egon.data.importing.industrial_sites as industrial_sites
-import egon.data.importing.mastr as mastr
+
 import egon.data.importing.nep_input_data as nep_input
-import egon.data.importing.openstreetmap as import_osm
-import egon.data.importing.re_potential_areas as re_potential_areas
 import egon.data.importing.scenarios as import_scenarios
-import egon.data.importing.vg250 as import_vg250
 import egon.data.importing.zensus as import_zs
+import egon.data.importing.gas_grid as gas_grid
+
+import egon.data.processing.boundaries_grid_districts as boundaries_grid_districts
 import egon.data.processing.demandregio as process_dr
 import egon.data.processing.district_heating_areas as district_heating_areas
-import egon.data.processing.loadarea as loadarea
-import egon.data.processing.openstreetmap as process_osm
 import egon.data.processing.osmtgmod as osmtgmod
 import egon.data.processing.power_plants as power_plants
 import egon.data.processing.renewable_feedin as import_feedin
 import egon.data.processing.substation as substation
 import egon.data.processing.zensus_vg250.zensus_population_inside_germany as zensus_vg250
-import egon.data.processing.mv_grid_districts as mvgd
+import egon.data.processing.gas_areas as gas_areas
+import egon.data.processing.wind_farms as wf
+import egon.data.processing.pv_ground_mounted as pv_gm
+import egon.data.importing.scenarios as import_scenarios
+import egon.data.importing.industrial_sites as industrial_sites
+import egon.data.processing.loadarea as loadarea
+import egon.data.processing.calculate_dlr as dlr
+
 import egon.data.processing.zensus as process_zs
 import egon.data.processing.zensus_grid_districts as zensus_grid_districts
 import egon.data.processing.heat_demand_timeseries.HTS as heat_time_series
+
 
 from egon.data import db
 
@@ -54,55 +68,26 @@ with airflow.DAG(
     is_paused_upon_creation=False,
     schedule_interval=None,
 ) as pipeline:
-    setup = PythonOperator(task_id="initdb", python_callable=initdb)
 
-    # Openstreetmap data import
-    osm_download = PythonOperator(
-        task_id="download-osm",
-        python_callable=import_osm.download_pbf_file,
-    )
-    osm_import = PythonOperator(
-        task_id="import-osm",
-        python_callable=import_osm.to_postgres,
-    )
-    osm_migrate = PythonOperator(
-        task_id="migrate-osm",
-        python_callable=process_osm.modify_tables,
-    )
-    osm_add_metadata = PythonOperator(
-        task_id="add-osm-metadata",
-        python_callable=import_osm.add_metadata,
-    )
-    setup >> osm_download >> osm_import >> osm_migrate >> osm_add_metadata
+    tasks = pipeline.task_dict
+
+    database_setup = database.Setup()
+    database_setup.insert_into(pipeline)
+    setup = tasks["database.setup"]
+
+    osm = OpenStreetMap(dependencies=[setup])
+    osm.insert_into(pipeline)
+    osm_add_metadata = tasks["osm.add-metadata"]
+    osm_download = tasks["osm.download"]
+
+    data_bundle = DataBundle(dependencies=[setup])
+    data_bundle.insert_into(pipeline)
+    download_data_bundle = tasks["data_bundle.download"]
 
     # VG250 (Verwaltungsgebiete 250) data import
-    vg250_download = PythonOperator(
-        task_id="download-vg250",
-        python_callable=import_vg250.download_vg250_files,
-    )
-    vg250_import = PythonOperator(
-        task_id="import-vg250",
-        python_callable=import_vg250.to_postgres,
-    )
-
-    vg250_nuts_mview = PostgresOperator(
-        task_id="vg250_nuts_mview",
-        sql="vg250_lan_nuts_id_mview.sql",
-        postgres_conn_id="egon_data",
-        autocommit=True,
-    )
-    vg250_metadata = PythonOperator(
-        task_id="add-vg250-metadata",
-        python_callable=import_vg250.add_metadata,
-    )
-    vg250_clean_and_prepare = PostgresOperator(
-        task_id="vg250_clean_and_prepare",
-        sql="cleaning_and_preparation.sql",
-        postgres_conn_id="egon_data",
-        autocommit=True,
-    )
-    setup >> vg250_download >> vg250_import >> vg250_nuts_mview
-    vg250_nuts_mview >> vg250_metadata >> vg250_clean_and_prepare
+    vg250 = Vg250(dependencies=[setup])
+    vg250.insert_into(pipeline)
+    vg250_clean_and_prepare = tasks["vg250.cleaning-and-preperation"]
 
     # Zensus import
     zensus_download_population = PythonOperator(
@@ -222,6 +207,7 @@ with airflow.DAG(
     vg250_clean_and_prepare >> demandregio_demand_cts_ind
     demandregio_tables >> demandregio_demand_cts_ind
     scenario_input_import >> demandregio_demand_cts_ind
+    download_data_bundle >> demandregio_demand_cts_ind
 
     # Society prognosis
     prognosis_tables = PythonOperator(
@@ -281,20 +267,19 @@ with airflow.DAG(
     setup >> create_tables >> nep_insert_data
     vg250_clean_and_prepare >> nep_insert_data
     population_import >> nep_insert_data
+    download_data_bundle >> nep_insert_data
 
     # setting etrago input tables
     etrago_input_data = PythonOperator(
         task_id="setting-etrago-input-tables",
-        python_callable=etrago.create_tables,
+        python_callable=etrago.setup,
     )
     setup >> etrago_input_data
 
     # Retrieve MaStR data
-    retrieve_mastr_data = PythonOperator(
-        task_id="retrieve_mastr_data",
-        python_callable=mastr.download_mastr_data,
-    )
-    setup >> retrieve_mastr_data
+    mastr_data = mastr_data_setup(dependencies=[setup])
+    mastr_data.insert_into(pipeline)
+    retrieve_mastr_data = tasks["mastr.download-mastr-data"]
 
     # Substation extraction
     substation_tables = PythonOperator(
@@ -320,6 +305,7 @@ with airflow.DAG(
         postgres_conn_id="egon_data",
         autocommit=True,
     )
+
 
     osm_add_metadata >> substation_tables >> substation_functions
     substation_functions >> hvmv_substation_extraction
@@ -354,37 +340,25 @@ with airflow.DAG(
     ehv_substation_extraction >> run_osmtgmod
     hvmv_substation_extraction >> run_osmtgmod
     run_osmtgmod >> osmtgmod_pypsa
+    etrago_input_data >> osmtgmod_pypsa
     run_osmtgmod >> osmtgmod_substation
 
+    # create Voronoi for MV grid districts
+    create_voronoi_substation = PythonOperator(
+        task_id="create-voronoi-substations",
+        python_callable=substation.create_voronoi,
+    )
+    osmtgmod_substation >> create_voronoi_substation
+
     # MV grid districts
-    create_voronoi = PythonOperator(
-        task_id="create_voronoi",
-        python_callable=substation.create_voronoi
-    )
-    osmtgmod_substation >> create_voronoi
-
-
-    define_mv_grid_districts = PythonOperator(
-        task_id="define_mv_grid_districts",
-        python_callable=mvgd.define_mv_grid_districts
-    )
-    create_voronoi >> define_mv_grid_districts
+    mv_grid_districts = mv_grid_districts_setup(dependencies=[create_voronoi_substation])
+    mv_grid_districts.insert_into(pipeline)
+    define_mv_grid_districts = tasks["mv_grid_districts.define-mv-grid-districts"]
 
     # Import potential areas for wind onshore and ground-mounted PV
-    download_re_potential_areas = PythonOperator(
-        task_id="download_re_potential_area_data",
-        python_callable=re_potential_areas.download_datasets,
-    )
-    create_re_potential_areas_tables = PythonOperator(
-        task_id="create_re_potential_areas_tables",
-        python_callable=re_potential_areas.create_tables,
-    )
-    insert_re_potential_areas = PythonOperator(
-        task_id="insert_re_potential_areas",
-        python_callable=re_potential_areas.insert_data,
-    )
-    setup >> download_re_potential_areas >> create_re_potential_areas_tables
-    create_re_potential_areas_tables >> insert_re_potential_areas
+    re_potential_areas = re_potential_area_setup(dependencies=[setup])
+    re_potential_areas.insert_into(pipeline)
+    insert_re_potential_areas = tasks["re_potential_areas.insert-data"]
 
     # Future heat demand calculation based on Peta5_0_1 data
     heat_demand_import = PythonOperator(
@@ -442,6 +416,25 @@ with airflow.DAG(
     demandregio_demand_cts_ind >> elec_cts_demands_zensus
     map_zensus_vg250 >> elec_cts_demands_zensus
 
+
+    # Gas grid import
+    gas_grid_insert_data = PythonOperator(
+        task_id="insert-gas-grid",
+        python_callable=gas_grid.insert_gas_data,
+    )
+
+    etrago_input_data >> gas_grid_insert_data
+    download_data_bundle >> gas_grid_insert_data
+
+    # Create gas voronoi
+    create_gas_polygons = PythonOperator(
+        task_id="create-gas-voronoi",
+        python_callable=gas_areas.create_voronoi,
+    )
+
+    gas_grid_insert_data  >> create_gas_polygons
+    vg250_clean_and_prepare >> create_gas_polygons
+
     # Extract landuse areas from osm data set
     create_landuse_table = PythonOperator(
         task_id="create-landuse-table",
@@ -459,7 +452,38 @@ with airflow.DAG(
     osm_add_metadata >> landuse_extraction
     vg250_clean_and_prepare >> landuse_extraction
 
- # Import weather data
+    # Generate wind power farms
+    generate_wind_farms = PythonOperator(
+        task_id="generate_wind_farms",
+        python_callable=wf.wind_power_parks,
+    )
+    retrieve_mastr_data >> generate_wind_farms
+    insert_re_potential_areas >> generate_wind_farms
+    scenario_input_import >> generate_wind_farms
+    hvmv_substation_extraction >> generate_wind_farms
+    define_mv_grid_districts >> generate_wind_farms
+
+    # Regionalization of PV ground mounted
+    generate_pv_ground_mounted = PythonOperator(
+        task_id="generate_pv_ground_mounted",
+        python_callable=pv_gm.regio_of_pv_ground_mounted,
+    )
+    retrieve_mastr_data >> generate_pv_ground_mounted
+    insert_re_potential_areas >> generate_pv_ground_mounted
+    scenario_input_import >> generate_pv_ground_mounted
+    hvmv_substation_extraction >> generate_pv_ground_mounted
+    define_mv_grid_districts >> generate_pv_ground_mounted
+
+    # Calculate dynamic line rating for HV trans lines
+
+    calculate_dlr = PythonOperator(
+        task_id="calculate_dlr",
+        python_callable=dlr.Calculate_DLR,
+    )
+    osmtgmod_pypsa >> calculate_dlr
+    download_data_bundle >> calculate_dlr
+
+    # Import weather data
     download_era5 = PythonOperator(
         task_id="download-weather-data",
         python_callable=import_era5.download_era5,
@@ -533,15 +557,53 @@ with airflow.DAG(
     map_zensus_vg250 >> electrical_load_curves_cts
     etrago_input_data >> electrical_load_curves_cts
 
-    # Heat demand time series
-    heat_time_series = PythonOperator(
-         task_id="heat-time-series",
-         python_callable=heat_time_series.demand_profile_generator,    
+
+    # Map federal states to mv_grid_districts
+    map_boundaries_grid_districts = PythonOperator(
+        task_id="map_vg250_grid_districts",
+        python_callable=boundaries_grid_districts.map_mvgriddistricts_vg250,
     )
-    
-    heat_demand_import >> heat_time_series
-    import_district_heating_areas >> heat_time_series
-    import_district_heating_areas >> heat_time_series
-    vg250_population_metadata >> heat_time_series
-    map_zensus_grid_districts >> heat_time_series
-    
+    define_mv_grid_districts >> map_boundaries_grid_districts
+    vg250_clean_and_prepare >> map_boundaries_grid_districts
+
+    # Solar rooftop per mv grid district
+    solar_rooftop_etrago = PythonOperator(
+        task_id="etrago_solar_rooftop",
+        python_callable=power_plants.pv_rooftop_per_mv_grid,
+    )
+    map_boundaries_grid_districts >> solar_rooftop_etrago
+    feedin_pv >> solar_rooftop_etrago
+    elec_cts_demands_zensus >> solar_rooftop_etrago
+    elec_household_demands_zensus >> solar_rooftop_etrago
+    nep_insert_data >> solar_rooftop_etrago
+    etrago_input_data >> solar_rooftop_etrago
+    map_zensus_grid_districts >> solar_rooftop_etrago
+
+    # Heat supply
+    heat_supply = HeatSupply(
+        dependencies=[data_bundle])
+
+    import_district_heating_supply = tasks["heat_supply.district-heating"]
+    import_individual_heating_supply = tasks["heat_supply.individual-heating"]
+    heat_supply_tables = tasks["heat_supply.create-tables"]
+    geothermal_potential = tasks["heat_supply.geothermal.potential-germany"]
+
+    create_district_heating_areas_table >> heat_supply_tables
+    import_district_heating_areas >> import_district_heating_supply
+    map_zensus_grid_districts >> import_district_heating_supply
+    import_district_heating_areas >> geothermal_potential
+    import_district_heating_areas >> import_individual_heating_supply
+    map_zensus_grid_districts >> import_individual_heating_supply
+    power_plant_import >> import_individual_heating_supply
+
+    # Heat to eTraGo
+    heat_etrago = HeatEtrago(
+        dependencies=[heat_supply])
+
+    heat_etrago_buses = tasks["heat_etrago.buses"]
+    heat_etrago_supply = tasks["heat_etrago.supply"]
+
+    etrago_input_data >> heat_etrago_buses
+    define_mv_grid_districts >> heat_etrago_buses
+    import_district_heating_supply >> heat_etrago_supply
+