+++ src/egon/data/cli.py
@@ -121,6 +121,26 @@ import egon.data.config as config
     ),
     show_default=True,
 )
+
+@click.option(
+    "--compose-project-name",
+    default="egon-data",
+    metavar="PROJECT",
+    help=(
+        "The name of the Docker project."
+        " Different compose_project_names are needed to run multiple instances"
+        " of egon-data on the same machine."
+    ),
+    show_default=True,
+)
+
+@click.option(
+    "--airflow-port",
+    default=8080,
+    metavar="AIRFLOW_PORT",
+    help=("Specify the port on which airflow runs."),
+    show_default=True,
+)
 @click.version_option(version=egon.data.__version__)
 @click.pass_context
 def egon_data(context, **kwargs):
@@ -279,7 +299,8 @@ def egon_data(context, **kwargs):
         )
     if code != 0:
         subprocess.run(
-            ["docker-compose", "up", "-d", "--build"],
+            ["docker-compose", "-p", options["--compose-project-name"],
+             "up", "-d", "--build"],
             cwd=str((Path(".") / "docker").absolute()),
         )
         time.sleep(1.5)  # Give the container time to boot.
@@ -341,6 +362,12 @@ def egon_data(context, **kwargs):
     connection.schema = options["--database-name"]
     airflow.add(connection)
     airflow.commit()
+
+    # TODO: This should probably rather be done during the database
+    #       initialization workflow task.
+    from egon.data.datasets import setup
+
+    setup()
 
 
 @egon_data.command(